requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
PyYAML
PyMySQL
psycopg2-binary
